+++ CHANGELOG.md
@@ -6,6 +6,8 @@ section below it for the last release. -->
 
 * Reload `e3-testsuite` tests automatically after configuration changes
 * Provide completion for tool switches (e.g., `Default_Switches`) in GPR files
+* New refactoring: [Generate Package Body](https://github.com/AdaCore/ada_language_server/blob/master/doc/refactoring_tools.md#generate-package-body)
+* New refactoring: [Generate Subprogram Body](https://github.com/AdaCore/ada_language_server/blob/master/doc/refactoring_tools.md#generate-subprogram-body)
 
 ## 2026.0.202510141